@@ -3,12 +3,15 @@
   "description": "",
   "version": "1.0.0",
   "dependencies": {
-    "@chakra-ui/react": "^1.8.9",
+    "@chakra-ui/react": "2.0.0",
+    "@chakra-ui/toast": "2.0.0",
+    "@chakra-ui/styled-system": "2.6.0",
     "@cmsgov/design-system": "^5.0.2",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@hookform/resolvers": "^2.9.11",
-    "@tanstack/react-query": "4.0.5",
+    "@tanstack/react-query": "^4.36.1",
+    "@tanstack/react-query-devtools": "4",
     "@vitejs/plugin-react": "^4.3.1",
     "aws-amplify": "^6.6.4",
     "base-64": "^1.0.0",
@@ -21,8 +24,8 @@
     "node-fetch": "3.3.2",
     "node-forge": "1.3.1",
     "object-path": "^0.11.8",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "react-dropzone": "^11.7.1",
     "react-hook-form": "^7.53.0",
     "react-icons": "^4.12.0",
@@ -40,8 +43,8 @@
     "build": "tsc && vite build",
     "preview": "vite preview",
     "coverage": "jest --coverage",
-    "test": "jest --setupFiles dotenv/config",
-    "watchTest": "jest --setupFiles dotenv/config --watch",
+    "test": "jest",
+    "watchTest": "jest --watch",
     "postinstall": "patch-package"
   },
   "jest": {
@@ -129,15 +132,15 @@
     "@babel/preset-env": "^7.13.15",
     "@babel/preset-typescript": "^7.16.0",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^12.1.0",
+    "@testing-library/react": "13.0.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/base-64": "^1.0.0",
     "@types/file-saver": "^2.0.5",
     "@types/jest": "^27.4.0",
     "@types/jest-axe": "^3.5.4",
     "@types/object-path": "^0.11.1",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "18.0.0",
+    "@types/react-dom": "18.0.0",
     "@types/react-router-dom": "^5.3.2",
     "@types/uuid": "^8.3.3",
     "babel": "^6.23.0",
