@@ -1,4 +1,5 @@
 import * as CUI from "@chakra-ui/react";
+import { createStandaloneToast } from "@chakra-ui/toast";
 import * as QMR from "components";
 import { MonthPicker } from "components/MonthPicker";
 import { useFormContext } from "react-hook-form";
@@ -28,7 +29,7 @@ export const currentMonth = parseInt(format(new Date(), "M"));
 export const DateRangeError = ({ name }: { name: string }) => {
   const { setValue, watch } = useFormContext();
   const range = watch(name);
-  const toast = CUI.useToast();
+  const toast = createStandaloneToast();
   const startYear = parseInt(range?.startDate?.selectedYear);
   const startMonth = parseInt(range?.startDate?.selectedMonth);
 
